requirements.txt
Django
psycopg2-binary
numpy
